# fewer string formats